+++ Hash.cpp
@@ -13,7 +13,7 @@ uint64_t Hash(void* Arr, size_t ArrElemQuant, size_t ArrElemSize)
     for (size_t Arr_i = 0; Arr_i < ArrElemQuant * ArrElemSize; Arr_i++)
     {
         assert((ArrChar + Arr_i) != NULL);
-        ArrHash = (ArrHash * 33) ^ ArrChar[Arr_i];
+        ArrHash = (ArrHash * 33) ^ (uint64_t) ArrChar[Arr_i];
     }
 
     return ArrHash;
